:bento: refactored and reorganized code, added test for endpoint
