+++ anon_api/deps.py
@@ -1,0 +1,8 @@
+import logging
+
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.getLevelName('DEBUG'))
+logger.addHandler(logging.StreamHandler())
+
+QUEUES = {}
+API_VERSION = 3
