+++ anon_api/lib_graphql.py
@@ -1,7 +1,8 @@
 import graphene as g
-from anon_api.lib_misc import cfg_get, check_envs
+from .lib_cfg import (config)
+from .lib_misc import check_envs
 
-ALGO_CONFIG = cfg_get({})['algorithms']
+ALGO_CONFIG = config.key('algorithms')
 
 
 class Algorithm(g.ObjectType):
