+++ tests/behave/steps/api_steps.py
@@ -1,4 +1,5 @@
 import requests
+import json
 import os
 from behave import when, given, then
 
@@ -71,3 +72,48 @@ def step_impl(context):
     assert context.json['ref'] == context.ref
     assert context.json['status'] == 'text'
     assert isinstance(context.json['value'], str)
+
+
+@given(u'some example text')
+def step_impl(context):
+    context.txt = """
+Robert serait le fils d'Erlebert et le neveu de Robert, référendaire de Dagobert Ier.
+Il est connu dès 654 à la cour de Clovis II et l'historienne Ingrid Heidrich le dit maire du palais de Neustrie en 654, mais son avis n'est pas partagé, car la charge était alors tenue par Erchinoald.
+
+Il accéda ensuite aux charges de comte palatin puis de chancelier de Clotaire III, roi des Francs en Neustrie.
+
+Partisan d'Ébroïn, il fit exécuter sur son ordre et contre son gré saint Léger le 2 octobre 6776 (ou 6771, ou 6797).
+
+Il meurt peu après puisqu'un acte du 12 septembre 678, mentionne que sa veuve Théoda a hérité de ses biens, la veuve étant également décédée à la date de l'acte. Théoda serait un diminutif de Théodrade, et comme son fils Lambert fut placé sous la protection Théodard de Maastricht, évêque de Tongres et qu'il lui succède ensuite comme évêque, Christian Settipani propose de la voir comme une sœur de Théodard
+
+Il serait le grand-père de Lambert de Hesbaye et donc probablement un membre de la famille des Robertiens et un ancêtre direct des Capétiens.
+"""
+
+
+@when(u'I send text file for pseudonymisation')
+def step_impl(context):
+    payload = {
+        "_v": 1,
+        "_timestamp": 1239120938,
+        "algo_list": [{
+            "id": "anon_trazor",
+            "params": "{}"
+        }],
+        "format": "text",
+        "encoding": "utf8",
+        "text": context.txt
+    }
+    r = requests.post(
+        f'http://{context.api_host}:{context.api_port}/run',
+        json=payload
+    )
+    assert r.status_code == 200
+    context.json = r.json()
+
+
+@then(u'I receive a pseudonymised text')
+def step_impl(context):
+    parsed_text = context.json['text']
+    print(parsed_text)
+    assert('Robert' not in parsed_text)
+    assert('Lambert' not in parsed_text)
